Fixing tests + fix problem at initialization
